% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc_oath_hearing_case_status.R
\name{nyc_oath_hearing_case_status}
\alias{nyc_oath_hearing_case_status}
\title{OATH Hearings Division Case Status}
\source{
NYC Open Data: <https://data.cityofnewyork.us/City-Government/OATH-Hearings-Division-Case-Status/jz4z-kudi/about_data>
}
\usage{
nyc_oath_hearing_case_status(limit = 10000, filters = list())
}
\arguments{
\item{limit}{Number of rows to retrieve (default = 10,000).}

\item{filters}{Optional list of field-value pairs to filter results.}
}
\value{
A tibble containing OATH Hearings Division Case Status data.
}
\description{
Downloads OATH Hearings Division Case Status data from NYC Open Data.
}
\details{
The OATH Hearings Division Case Status dataset contains information about alleged public safety and quality of life violations that are filed and adjudicated through the City’s administrative law court, the NYC Office of Administrative Trials and Hearings (OATH) and provides information about the infraction charged, decision outcome, payments, amounts and fees relating to the case.
The summonses listed in this dataset are issued and filed at the OATH Hearings Division by City enforcement agencies.
}
\examples{
# Quick example (fetch 5 rows)
small_sample <- nyc_oath_hearing_case_status(limit = 5)
head(small_sample)

\donttest{
nyc_oath_hearing_case_status(limit = 5000)
nyc_oath_hearing_case_status(filters = list(respondent_address_borough = "BROOKLYN"))
}
}
