% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_creation.R
\name{nyc_311}
\alias{nyc_311}
\title{NYC 311 Service Requests}
\usage{
nyc_311(limit = 10000, filters = list())
}
\arguments{
\item{limit}{Number of rows to retrieve (default = 10,000).}

\item{filters}{Optional list of field-value pairs to filter results.}
}
\value{
A tibble containing 311 Service Request data.
}
\description{
Downloads NYC 311 Service Request data from NYC Open Data.
}
\examples{
\dontrun{
nyc_311(limit = 5000)
nyc_311(filters = list(agency = "NYPD", city = "BROOKLYN"))
}
}
