% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc_list_datasets.R
\name{nyc_list_datasets}
\alias{nyc_list_datasets}
\title{List datasets available in nycOpenData}
\usage{
nyc_list_datasets()
}
\value{
A tibble of available datasets.
}
\description{
Returns a catalog of datasets available from NYC Open Data.
}
\examples{
if (interactive() && curl::has_internet()) {
  nyc_list_datasets()
}
}
