% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc_museum_locations.R
\name{nyc_museum_locations}
\alias{nyc_museum_locations}
\title{MUSEUM}
\source{
NYC Open Data: <https://data.cityofnewyork.us/Recreation/MUSEUM/fn6f-htvy/about_data>
}
\usage{
nyc_museum_locations(limit = 10000, filters = list())
}
\arguments{
\item{limit}{Number of rows to retrieve (default = 10,000).}

\item{filters}{Optional list of field-value pairs to filter results.}
}
\value{
A tibble containing Locations of New York City Museums data.
}
\description{
Downloads Locations of New York City Museums from NYC Open Data.
}
\details{
Locations of New York City Museums
}
\examples{
# Quick example (fetch 10 rows)
small_sample <- nyc_museum_locations(limit = 10)
head(small_sample)

\donttest{
nyc_museum_locations(limit = 5000)
nyc_museum_locations(filters = list(city = "New York"))
}
}
